# 默认模块.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "gowvp/owl HTTP API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/events": {
      "get": {
        "summary": "事件列表",
        "tags": [
          "事件"
        ],
        "x-apifox-folder": "事件",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "页码，从 1 开始"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "description": "每页数量"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "排序字段，支持前缀 - 表示倒序"
            }
          },
          {
            "name": "did",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "label",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_ms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "end_ms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/events/{id}": {
      "get": {
        "summary": "事件详情",
        "tags": [
          "事件"
        ],
        "x-apifox-folder": "事件",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "更新事件",
        "tags": [
          "事件"
        ],
        "x-apifox-folder": "事件",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ended_at": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "删除事件",
        "tags": [
          "事件"
        ],
        "x-apifox-folder": "事件",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/webhook/events": {
      "post": {
        "summary": "接收告警事件",
        "tags": [
          "Webhook"
        ],
        "x-apifox-folder": "Webhook",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookForward"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultOutput"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/ai/events": {
      "post": {
        "summary": "AI 检测事件",
        "tags": [
          "Webhook"
        ],
        "x-apifox-folder": "Webhook",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIDetectionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultOutput"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DefaultOutput": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          }
        }
      },
...
      "Event": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "did": {
            "type": "string"
          },
          "cid": {
            "type": "string"
          },
          "started_at": {
            "type": "string",
            "format": "date-time"
          },
          "ended_at": {
            "type": "string",
            "format": "date-time"
          },
          "label": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "zones": {
            "type": "string"
          },
          "image_path": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AIBoundingBox": {
        "type": "object",
        "properties": {
          "x_min": {
            "type": "integer",
            "format": "int32"
          },
          "y_min": {
            "type": "integer",
            "format": "int32"
          },
          "x_max": {
            "type": "integer",
            "format": "int32"
          },
          "y_max": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AINormBox": {
        "type": "object",
        "properties": {
          "x": {
            "type": "number",
            "format": "double"
          },
          "y": {
            "type": "number",
            "format": "double"
          },
          "w": {
            "type": "number",
            "format": "double"
          },
          "h": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "AIDetection": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "confidence": {
            "type": "number",
            "format": "double"
          },
          "box": {
            "$ref": "#/components/schemas/AIBoundingBox"
          },
          "area": {
            "type": "integer",
            "format": "int32"
          },
          "norm_box": {
            "$ref": "#/components/schemas/AINormBox"
          }
        }
      },
      "AIDetectionInput": {
        "type": "object",
        "properties": {
          "camera_id": {
            "type": "string"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          },
          "detections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AIDetection"
            }
          },
          "snapshot": {
            "type": "string"
          },
          "snapshot_width": {
            "type": "integer",
            "format": "int32"
          },
          "snapshot_height": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "WebhookForward": {
        "type": "object",
        "properties": {
          "did": {
            "type": "string"
          },
          "cid": {
            "type": "string"
          },
          "started_at": {
            "type": "integer",
            "format": "int64"
          },
          "ended_at": {
            "type": "integer",
            "format": "int64"
          },
          "label": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "zones": {
            "type": "string"
          },
          "image_base64": {
            "type": "string"
          },
          "image_path": {
            "type": "string"
          },
          "model": {
            "type": "string"
          }
        }
      },
...
      "EventList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "reason": {
                  "type": "string"
                },
                "msg": {
                  "type": "string"
                },
                "details": {
                  "type": "object",
                  "properties": {}
                },
                "trace_id": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}